+++ src/main/java/com/danvinicius/ecommerce/services/CartService.java
@@ -5,6 +5,7 @@ import java.util.List;
 import java.util.UUID;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import com.danvinicius.ecommerce.dto.cart.CartItemRequestDTO;
@@ -12,10 +13,14 @@ import com.danvinicius.ecommerce.dto.cart.CartRequestDTO;
 import com.danvinicius.ecommerce.entities.cart.Cart;
 import com.danvinicius.ecommerce.entities.cart.CartItem;
 import com.danvinicius.ecommerce.entities.product.Product;
+import com.danvinicius.ecommerce.entities.user.User;
+import com.danvinicius.ecommerce.entities.user.UserRole;
+import com.danvinicius.ecommerce.exceptions.ForbiddenException;
 import com.danvinicius.ecommerce.exceptions.ProductAlreadyInCartException;
 import com.danvinicius.ecommerce.exceptions.ProductNotInCartException;
-import com.danvinicius.ecommerce.exceptions.ProductUnavailable;
+import com.danvinicius.ecommerce.exceptions.ProductUnavailableException;
 import com.danvinicius.ecommerce.exceptions.ResourceNotFoundException;
+import com.danvinicius.ecommerce.repositories.CartItemRepository;
 import com.danvinicius.ecommerce.repositories.CartRepository;
 
 @Service
@@ -24,23 +29,53 @@ public class CartService {
     private CartRepository cartRepository;
 
     @Autowired
+    private CartItemRepository cartItemRepository;
+
+    @Autowired
     private ProductService productService;
 
+    @Autowired
+    private UserService userService;
+
+    private User user() {
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        return this.userService.getUserById(user.getId().toString());
+    }
+    
+    private Boolean userHasPermission(Cart cart) {
+        return this.user().getRole().equals(UserRole.ADMIN.getRole()) || cart.getUser().equals(this.user());
+    }
+
     public Cart getCartById(String id) throws ResourceNotFoundException {
-        return cartRepository.findById(UUID.fromString(id)).orElseThrow(ResourceNotFoundException::new);
+        Cart cart = cartRepository.findById(UUID.fromString(id)).orElseThrow(ResourceNotFoundException::new);
+        if (!userHasPermission(cart)) {
+            throw new ForbiddenException();
+        };
+        return cart;
     }
 
     public List<Cart> getAllCarts() {
         return cartRepository.findAll();
     }
 
+    public Cart emptyCart(Cart cart) {
+        for (CartItem cartItem : cart.getItems()) {
+            cartItemRepository.delete(cartItem);
+        }
+        cart.getItems().clear();
+        cart.setTotalPrice(BigDecimal.ZERO);
+        cartRepository.save(cart);
+        return cart;
+    }
+
     public Cart createCart(CartRequestDTO data) {
         Cart cart = new Cart();
+        cart.setUser(this.user());
         data.items().stream().forEach(item -> {
             Product product = productService.getProductById(item.productId());
             CartItem cartItem = new CartItem(null, cart, product, item.quantity(), product.getPrice());
             if (product.getQuantity() == 0 || item.quantity() > product.getQuantity()) {
-                throw new ProductUnavailable("Item quantity not available");
+                throw new ProductUnavailableException("Item quantity not available");
             }
             cart.getItems().add(cartItem);
             product.setQuantity(product.getQuantity() - item.quantity());
@@ -51,12 +86,17 @@ public class CartService {
         return cart;
     }
     
-    public Cart addItemToCart(String cartId, CartItemRequestDTO item) throws ProductAlreadyInCartException, ProductUnavailable {
+    public Cart addItemToCart(String cartId, CartItemRequestDTO item) throws ProductAlreadyInCartException, ProductUnavailableException {
         Cart cart = getCartById(cartId);
+
+        if (!userHasPermission(cart)) {
+            throw new ForbiddenException();
+        };
+
         Product product = productService.getProductById(item.productId());
 
         if (product.getQuantity() == 0 || item.quantity() > product.getQuantity()) {
-            throw new ProductUnavailable("Item quantity not available");
+            throw new ProductUnavailableException("Item quantity not available");
         }
 
         if (isProductInCart(cart, product)) {
@@ -81,6 +121,11 @@ public class CartService {
     
     public Cart removeItemFromCart(String cartId, CartItemRequestDTO item) throws ProductNotInCartException {
         Cart cart = getCartById(cartId);
+
+        if (!userHasPermission(cart)) {
+            throw new ForbiddenException();
+        };
+
         Product product = productService.getProductById(item.productId());
         if (!isProductInCart(cart, product)) {
             throw new ProductNotInCartException();
@@ -105,6 +150,9 @@ public class CartService {
     
     public void deleteCart(String id) {
         Cart cart = getCartById(id);
+        if (!userHasPermission(cart)) {
+            throw new ForbiddenException();
+        };
         cartRepository.delete(cart);
     }
